orchestrator/requirements.txt
google-cloud-storage
numpy
requests